+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - Extend the experimental setting
 
+## Version 1.1.2: July 26, 2021
+
+- Updated run task for windows such that the executable name has .exe file extension, since this is needed for MinGW
+- Fixed bug that debugging the release build was not possible
+
 ## Version 1.1.1: July 26, 2021
 
 - Creating and deleting build folder is now executed by the extension code and not anymore by the Makefile